requirements.txt
# PI0.5 Continual Learning Dependencies
# Pin versions to avoid conflicts between OpenPI (Dec 2025) and LIBERO (Mar 2025)

# Core ML
torch>=2.1.0,<2.3.0
torchvision>=0.16.0
transformers>=4.36.0,<4.40.0
accelerate>=0.25.0

# VLA / Robotics
lerobot>=0.1.0
# openpi - install from git: pip install git+https://github.com/Physical-Intelligence/openpi.git

# LIBERO simulation
# libero - install from git: pip install git+https://github.com/Lifelong-Robot-Learning/LIBERO.git
robosuite>=1.4.0
mujoco>=3.0.0

# LoRA / Fine-tuning
peft>=0.7.0
bitsandbytes>=0.41.0  # For 8-bit loading if OOM

# Data handling
datasets>=2.14.0
huggingface-hub>=0.19.0
h5py>=3.9.0
numpy>=1.24.0,<2.0.0

# Monitoring / Visualization
wandb>=0.16.0
matplotlib>=3.7.0
tensorboard>=2.14.0
tqdm>=4.66.0

# Utils
pyyaml>=6.0
omegaconf>=2.3.0
einops>=0.7.0
scipy>=1.11.0
pillow>=10.0.0

# Optional: for dashboard
# streamlit>=1.28.0
